Even moooore cleanup
